# docs/swagger.yaml
openapi: 3.0.4
info:
  title: PPO Project.HttpServer
  version: v1
paths:
...
  '/api/v1/scores/{scoreId}':
    get:
      tags:
        - Score
      parameters:
        - name: scoreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ScoreDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ScoreDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    delete:
      tags:
        - Score
      parameters:
        - name: scoreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    patch:
      tags:
        - Score
      parameters:
        - name: scoreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScoreDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateScoreDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateScoreDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ScoreDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ScoreDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /api/v1/scores:
    post:
      tags:
        - Score
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScoreDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateScoreDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateScoreDto'
        required: true
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ScoreDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ScoreDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  '/api/v1/employees/{employeeId}/scores':
    get:
      tags:
        - Score
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 1
            nullable: false
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            enum: [12, 24, 36]
            default: 12
            nullable: false
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  '/api/v1/employees/scoreAuthor/{authorId}/scores':
    get:
      tags:
        - Score
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 1
            nullable: false
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            enum: [12, 24, 36]
            default: 12
            nullable: false
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  '/api/v1/positions/{positionId}/scores':
    get:
      tags:
        - Score
      parameters:
        - name: positionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 1
            nullable: false
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            enum: [12, 24, 36]
            default: 12
            nullable: false
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  '/api/v1/employees/{headEmployeeId}/subordinates/scores':
    get:
      tags:
        - Score
      parameters:
        - name: headEmployeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 1
            nullable: false
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            enum: [12, 24, 36]
            default: 12
            nullable: false
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
components:
  schemas:
...
    CreateScoreDto:
      required:
        - employeeId
        - authorId
        - positionId
        - createdAt
        - efficiencyScore
        - engagementScore
        - competencyScore
      type: object
      properties:
        employeeId:
          type: string
          format: uuid
          nullable: false
        authorId:
          type: string
          format: uuid
          nullable: false
        positionId:
          type: string
          format: uuid
          nullable: false
        createdAt:
          type: string
          format: date-time
          nullable: false
        efficiencyScore:
          type: integer
          format: int32
          nullable: false
        engagementScore:
          type: integer
          format: int32
          nullable: false
        competencyScore:
          type: integer
          format: int32
          nullable: false
      additionalProperties: false
...
    ErrorDto:
      required:
        - errorType
        - message
      type: object
      properties:
        errorType:
          type: string
          nullable: false
        message:
          type: string
          nullable: true
      additionalProperties: false
...
    ScoreDto:
      required:
        - id
        - employeeId
        - authorId
        - positionId
        - createdAt
        - efficiencyScore
        - engagementScore
        - competencyScore
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        employeeId:
          type: string
          format: uuid
          nullable: false
        authorId:
          type: string
          format: uuid
          nullable: false
        positionId:
          type: string
          format: uuid
          nullable: false
        createdAt:
          type: string
          format: date-time
          nullable: false
        efficiencyScore:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
          nullable: false
        engagementScore:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
          nullable: false
        competencyScore:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
          nullable: false
      additionalProperties: false
...
    UpdateScoreDto:
      required:
        - createdAt
        - efficiencyScore
        - engagementScore
        - competencyScore
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
          nullable: true
        efficiencyScore:
          type: integer
          format: int32
          nullable: true
        engagementScore:
          type: integer
          format: int32
          nullable: true
        competencyScore:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
security:
  - Bearer: [ ]